# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Proof-Carrying Data Analyst API
  version: 1.0.0
  description: Stable MVP boundary for dataset ingestion, asynchronous analysis, verification, and evidence.
servers:
  - url: /api/v1
...
paths:
...
  /datasets:
    post:
      tags: [Datasets]
      operationId: createDataset
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                name: {type: string, minLength: 1, maxLength: 100}
                files:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items: {type: string, format: binary}
      responses:
        '202':
          description: Upload accepted and profiling started.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Dataset'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /datasets/{dataset_id}:
    parameters:
      - $ref: '#/components/parameters/DatasetId'
    get:
      tags: [Datasets]
      operationId: getDataset
      responses:
        '200':
          description: Dataset metadata and current status.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Dataset'}
        '404': {$ref: '#/components/responses/NotFound'}
  /datasets/{dataset_id}/profile:
    parameters:
      - $ref: '#/components/parameters/DatasetId'
    get:
      tags: [Datasets]
      operationId: getDatasetProfile
      responses:
        '200':
          description: Completed profile.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DatasetProfile'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /analyses:
    post:
      tags: [Analyses]
      operationId: createAnalysis
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateAnalysisRequest'}
      responses:
        '202':
          description: Analysis queued.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Analysis'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /analyses/{analysis_id}:
    parameters:
      - $ref: '#/components/parameters/AnalysisId'
    get:
      tags: [Analyses]
      operationId: getAnalysis
      responses:
        '200':
          description: Current analysis state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Analysis'}
        '404': {$ref: '#/components/responses/NotFound'}
  /analyses/{analysis_id}/events:
    parameters:
      - $ref: '#/components/parameters/AnalysisId'
      - name: Last-Event-ID
        in: header
        required: false
        schema: {type: string}
    get:
      tags: [Analyses]
      operationId: streamAnalysisEvents
      responses:
        '200':
          description: SSE stream. Events are snapshot, stage.changed, clarification.required, analysis.terminal, and heartbeat.
          content:
            text/event-stream:
              schema: {type: string}
        '404': {$ref: '#/components/responses/NotFound'}
  /analyses/{analysis_id}/clarifications:
    parameters:
      - $ref: '#/components/parameters/AnalysisId'
    post:
      tags: [Analyses]
      operationId: answerClarification
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ClarificationAnswer'}
      responses:
        '202':
          description: Clarification accepted; analysis resumed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Analysis'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /analyses/{analysis_id}/result:
    parameters:
      - $ref: '#/components/parameters/AnalysisId'
    get:
      tags: [Analyses]
      operationId: getAnalysisResult
      responses:
        '200':
          description: Terminal analysis result.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AnalysisResult'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /analyses/{analysis_id}/evidence:
    parameters:
      - $ref: '#/components/parameters/AnalysisId'
    get:
      tags: [Analyses]
      operationId: getAnalysisEvidence
      responses:
        '200':
          description: Immutable evidence manifest.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EvidencePackage'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /analyses/{analysis_id}/reruns:
    parameters:
      - $ref: '#/components/parameters/AnalysisId'
    post:
      tags: [Analyses]
      operationId: rerunAnalysis
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                question: {type: string, minLength: 3, maxLength: 4000}
      responses:
        '202':
          description: New immutable analysis queued.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Analysis'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /analyses/{analysis_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/AnalysisId'
    post:
      tags: [Analyses]
      operationId: cancelAnalysis
      responses:
        '202':
          description: Cancellation accepted or analysis already cancelled.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Analysis'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /analyses/{analysis_id}/recovery:
    post:
      operationId: recoverAnalysis
      summary: Supply a missing exchange rate and create a new verified analysis
      parameters:
        - {name: analysis_id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecoveryAnswer'}
      responses:
        '202':
          description: Recovery analysis queued
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Analysis'}
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 8, maxLength: 128}
    DatasetId:
      name: dataset_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    AnalysisId:
      name: analysis_id
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    NotFound:
      description: Resource not found or expired.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Conflict:
      description: Resource state conflicts with this operation.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    RequestTooLarge:
      description: A file or bundle exceeds effective limits.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    UnsupportedMediaType:
      description: One or more file types are unsupported.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    ValidationError:
      description: Request fields are invalid.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
  schemas:
...
    DatasetStatus:
      type: string
      enum: [uploading, profiling, ready, failed]
    DatasetFile:
      type: object
      additionalProperties: false
      required: [id, filename, media_type, size_bytes, sha256]
      properties:
        id: {type: string, format: uuid}
        filename: {type: string}
        media_type: {type: string}
        size_bytes: {type: integer, minimum: 0}
        sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
    Warning:
      type: object
      additionalProperties: false
      required: [code, message, severity]
      properties:
        code: {type: string}
        message: {type: string}
        severity: {type: string, enum: [info, warning, critical]}
        table_id: {type: [string, 'null']}
        column: {type: [string, 'null']}
    Dataset:
      type: object
      additionalProperties: false
      required: [id, name, status, created_at, expires_at, files, table_count, warnings]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        status: {$ref: '#/components/schemas/DatasetStatus'}
        created_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        files: {type: array, items: {$ref: '#/components/schemas/DatasetFile'}}
        table_count: {type: [integer, 'null'], minimum: 0}
        warnings: {type: array, items: {$ref: '#/components/schemas/Warning'}}
        failure: {anyOf: [{$ref: '#/components/schemas/Failure'}, {type: 'null'}]}
    DatasetProfile:
      type: object
      additionalProperties: false
      required: [dataset_id, generated_at, tables, join_candidates, warnings]
      properties:
        dataset_id: {type: string, format: uuid}
        generated_at: {type: string, format: date-time}
        tables: {type: array, items: {$ref: '#/components/schemas/TableProfile'}}
        join_candidates: {type: array, items: {$ref: '#/components/schemas/JoinCandidate'}}
        warnings: {type: array, items: {$ref: '#/components/schemas/Warning'}}
...
    CreateAnalysisRequest:
      type: object
      additionalProperties: false
      required: [dataset_id, question]
      properties:
        dataset_id: {type: string, format: uuid}
        question: {type: string, minLength: 3, maxLength: 4000}
        preferences:
          type: object
          additionalProperties: false
          properties:
            answer_format: {type: string, enum: [auto, text, table, chart], default: auto}
            allow_explicit_assumptions: {type: boolean, default: true}
...
    Outcome:
      type: [string, 'null']
      enum: [answered, partial, refused, failed, cancelled, null]
...
    Analysis:
      type: object
      additionalProperties: false
      required: [id, dataset_id, parent_analysis_id, status, outcome, stage, created_at, updated_at, clarification, failure]
      properties:
        id: {type: string, format: uuid}
        dataset_id: {type: string, format: uuid}
        parent_analysis_id: {type: [string, 'null'], format: uuid}
        status: {$ref: '#/components/schemas/AnalysisStatus'}
        outcome: {$ref: '#/components/schemas/Outcome'}
        stage: {$ref: '#/components/schemas/Stage'}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        clarification: {anyOf: [{$ref: '#/components/schemas/Clarification'}, {type: 'null'}]}
        failure: {anyOf: [{$ref: '#/components/schemas/Failure'}, {type: 'null'}]}
...
    ClarificationAnswer:
      type: object
      additionalProperties: false
      required: [clarification_id]
      properties:
        clarification_id: {type: string}
        choice_id: {type: string}
        text: {type: string, minLength: 1, maxLength: 2000}
      oneOf:
        - required: [choice_id]
          not: {required: [text]}
        - required: [text]
          not: {required: [choice_id]}
    Failure:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
    AnalysisResult:
      type: object
      additionalProperties: false
      required: [analysis_id, outcome, headline, narrative, metrics, tables, charts, assumptions, warnings, confidence, verification, evidence_url]
      properties:
        analysis_id: {type: string, format: uuid}
        outcome: {type: string, enum: [answered, partial, refused, failed, cancelled]}
        headline: {type: [string, 'null']}
        narrative: {type: [string, 'null']}
        metrics: {type: array, items: {$ref: '#/components/schemas/MetricResult'}}
        tables: {type: array, items: {$ref: '#/components/schemas/TableResult'}}
        charts: {type: array, items: {$ref: '#/components/schemas/ChartResult'}}
        assumptions: {type: array, items: {$ref: '#/components/schemas/Assumption'}}
        warnings: {type: array, items: {$ref: '#/components/schemas/Warning'}}
        confidence: {$ref: '#/components/schemas/Confidence'}
        verification: {$ref: '#/components/schemas/VerificationSummary'}
        evidence_url: {type: [string, 'null']}
        review: {$ref: '#/components/schemas/AnswerReview'}
        recovery: {$ref: '#/components/schemas/AnswerRecovery'}
...
    EvidencePackage:
      type: object
      additionalProperties: false
      required: [analysis_id, created_at, plan, code_artifacts, claims, checks, execution]
      properties:
        analysis_id: {type: string, format: uuid}
        created_at: {type: string, format: date-time}
        plan: {$ref: '#/components/schemas/AnalysisPlan'}
        code_artifacts: {type: array, items: {$ref: '#/components/schemas/CodeArtifact'}}
        claims: {type: array, items: {$ref: '#/components/schemas/EvidenceClaim'}}
        checks: {type: array, items: {$ref: '#/components/schemas/VerificationCheck'}}
        execution: {$ref: '#/components/schemas/ExecutionRecord'}
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id, retryable, details]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}
            retryable: {type: boolean}
            details: {type: object, additionalProperties: true}
    RecoveryAnswer:
      type: object
      additionalProperties: false
      required: [rate, rate_source]
      properties:
        rate: {type: number, exclusiveMinimum: 0, maximum: 1000000}
        rate_source: {type: string, minLength: 3, maxLength: 200}
...